refactor: change inputs directory
