+++ src/components/basicComponents/Inputs/RadioInput.tsx
@@ -1,7 +1,7 @@
 import { FieldErrors, FieldValues, Path, UseFormRegister, WatchInternal } from 'react-hook-form';
-import { IFormValues } from '../CatsPageComponent/CatsPage';
+import { IFormValues } from '../../CatsPageComponent/CatsPage';
 import './RadioInput.css';
-import { Tooltip } from './Tooltip';
+import { Tooltip } from '../Tooltip';
 
 interface RadioInputProps {
   label: Path<IFormValues>;
