+++ src/components/basicComponents/Inputs/SelectInput.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent, useState } from 'react';
 import { FieldErrors, FieldValues, Path, UseFormRegister } from 'react-hook-form';
-import { IFormValues } from '../CatsPageComponent/CatsPage';
+import { IFormValues } from '../../CatsPageComponent/CatsPage';
 import './SelectInput.css';
-import { Tooltip } from './Tooltip';
+import { Tooltip } from '../Tooltip';
 
 interface SelectInputProps {
   label: Path<IFormValues>;
